bytecode events

+++ barretenberg/cpp/src/barretenberg/vm2/simulation/testing/mock_bytecode_manager.hpp
@@ -16,7 +16,7 @@ class MockBytecodeManager : public BytecodeManagerInterface {
     MockBytecodeManager();
     ~MockBytecodeManager() override;
 
-    MOCK_METHOD((std::pair<Instruction, /*read_bytes*/ uint32_t>), read_instruction, (size_t pc), (const, override));
+    MOCK_METHOD((std::pair<Instruction, /*read_bytes*/ uint32_t>), read_instruction, (uint32_t pc), (const, override));
     MOCK_METHOD(ContractClassId, get_class_id, (), (const, override));
 };
